refactor: removed ConferencingView component and migrated its functionality to ConferencingAppsViewWebWrapper
